chore: increase coverage
