+++ tests/resources/test_cameras.py
@@ -1,73 +1,3 @@
-from collections import namedtuple
-from datetime import datetime
-from string import Template
-from botocore.exceptions import ClientError
-from pinthesky import api
-from pinthesky.globals import app_context
-from unittest.mock import MagicMock
-import json
-import pytest
-
-
-Response = namedtuple('Response', field_names=['code', 'body', 'headers'])
-
-
-class Resources():
-    def __init__(self, module) -> None:
-        self.module = module
-        self.resource_name = module.__name__.split('.')[-1]
-
-    def __call__(self, *args, **kwds):
-        return self.request('/'.join(args), **kwds)
-
-    def request(self, name="", method='GET', query_params={}, body=None):
-        path = f'/{self.resource_name}{name}'
-        with open('events/resources/request.template.json') as f:
-            content = f.read()
-            template = Template(content)
-            event = json.loads(template.safe_substitute(
-                path=path,
-                method=method,
-                body=json.dumps(json.dumps(body)) if body is not None else '""'
-            ))
-            event['queryStringParameters'] = query_params
-        res = api(event=event, context={})
-        return Response(
-            code=res['statusCode'],
-            body=json.loads(res['body']) if res['body'] is not None else "",
-            headers=res['headers']
-        )
-
-
-@pytest.fixture(scope="module")
-def cameras(table):
-    assert table.name == 'Pits'
-    from pinthesky.resource import cameras
-
-    iot = MagicMock()
-    iot_data = MagicMock()
-
-    def side_effect(Bucket, Key):
-        if Key == 'images/PitsCamera1/thumbnail_latest.jpg':
-            return {
-                'ContentType': 'text/plain',
-                'ContentLength': 100,
-                'LastModified': datetime.now()
-            }
-        else:
-            raise ClientError({
-                'Error': {
-                    'Code': '404'
-                }
-            }, 'head_object')
-    s3 = MagicMock()
-    s3.head_object = MagicMock()
-    s3.head_object.side_effect = side_effect
-    app_context.inject('iot', iot, force=True)
-    app_context.inject('iot_data', iot_data, force=True)
-    app_context.inject('s3', s3, force=True)
-    return Resources(cameras)
-
 
 def test_camera_crud_workflow(cameras):
     # List, confirm empty
